Revert "use sp to avoid bcast error"
This reverts commit a1b23fab41dc46d93008358c37b000fd0eb82209.

+++ src/units.f90
@@ -29,7 +29,7 @@ contains
       logical, optional                   :: bcast
       logical                             :: bcast_loc
       type(CFG_t)                         :: cfg
-      real(sp)                             :: factor
+      real(dp)                             :: factor
       integer                             :: ierr
       character(len=300)                  :: unit_name
